+++ test/cases/html-plugin/webpack.config.js
@@ -4,14 +4,17 @@ import HtmlWebpackPlugin from 'html-webpack-plugin';
 module.exports = {
     entry: './index.js',
     plugins: [
-        new HtmlWebpackPlugin({ template: './index.ejs' }),
+        new HtmlWebpackPlugin({
+            template: './index.ejs',
+            minify: false,
+            scriptLoading: 'blocking',
+        }),
         new ConcatPlugin({
-            uglify: true,
-            sourceMap: true,
             name: 'file',
             fileName: '[name].[hash:20].js',
             injectType: 'none',
             filesToConcat: ['../../fixtures/a.js', '../../fixtures/b.js']
         }),
     ],
+    devtool: 'source-map',
 };