+++ test/cases/source-map-use-hash/expected/file.0ff883d05f62ecabac00.js
@@ -6,7 +6,6 @@ var x = 1;
  * @type {number}
  */
 x = 10;
-
 var y = 2;
 
 
@@ -18,4 +17,5 @@ var y = 2;
 y = 5;
 debugger;
 // End of this test file
+
 //# sourceMappingURL=file.0ff883d05f62ecabac00.js.map