+++ test/cases/use-hash/expected/file.0ff883d05f62ecabac00.js
@@ -6,7 +6,6 @@ var x = 1;
  * @type {number}
  */
 x = 10;
-
 var y = 2;